requirements.txt
python-telegram-bot
torch
transformers
tabulate